pyproject.toml
[project]
name = "openmeteo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "mysql-connector>=2.2.9",
    "pathlib>=1.0.1",
    "pymysql>=1.1.2",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "urllib3>=2.0.0",
    "ipython>=9.8.0",
]
